--- data/fs/files/product.json
@@ -1,143 +1,0 @@
-[
-  {
-    "id": "335ec12bef2341daecdd567c",
-    "title": "Gorra Trucker",
-    "photo": "https://picsum.photos/",
-    "category": "accesorios",
-    "price": 14800,
-    "stock": 9
-  },
-  {
-    "id": "6d75a713ee78c5db41777a5d",
-    "title": "Pulcera Cuero",
-    "photo": "https://picsum.photos/",
-    "category": "accesorios",
-    "price": 8800,
-    "stock": 12
-  },
-  {
-    "id": "719471191b11fbb44a59984b",
-    "title": "Lentes Sol",
-    "photo": "https://picsum.photos/",
-    "category": "accesorios",
-    "price": 19800,
-    "stock": 2
-  },
-  {
-    "id": "ec4d94600768e4691d1b5a1c",
-    "title": "Zapas Vans",
-    "photo": "https://picsum.photos/",
-    "category": "calzado",
-    "price": 48500,
-    "stock": 5
-  },
-  {
-    "id": "e1a4f65a26424c9d86acda25",
-    "title": "Zapas Adidas",
-    "photo": "https://picsum.photos/",
-    "category": "calzado",
-    "price": 95000,
-    "stock": 3
-  },
-  {
-    "id": "64f27e141793e83d50553ee6",
-    "title": "Zapas Nike",
-    "photo": "https://picsum.photos/",
-    "category": "calzado",
-    "price": 103200,
-    "stock": 1
-  },
-  {
-    "id": "5d1b09690e90b1b7f0d19bda",
-    "title": "Zapas Puma",
-    "photo": "https://picsum.photos/",
-    "category": "calzado",
-    "price": 79000,
-    "stock": 8
-  },
-  {
-    "id": "28f7b9d927d4ccaa06c3c3b0",
-    "title": "Zapas Topper",
-    "photo": "https://picsum.photos/",
-    "category": "calzado",
-    "price": 12000,
-    "stock": 13
-  },
-  {
-    "id": "35a6e416eef5df83bfe02355",
-    "title": "Pulsera Hombre",
-    "photo": "https://picsum.photos/",
-    "category": "accesorios",
-    "price": 3000,
-    "stock": 17
-  },
-  {
-    "id": "046e6348ab3d78b16e2a2c82",
-    "title": "Riñonera",
-    "photo": "https://picsum.photos/",
-    "category": "accesorios",
-    "price": 20600,
-    "stock": 1
-  },
-  {
-    "id": "3eedfa8ac9ab3fbc9921453c",
-    "title": "Gorra Dama",
-    "photo": "https://picsum.photos/",
-    "category": "accesorios",
-    "price": 14850,
-    "stock": 8
-  },
-  {
-    "id": "1afedf45c20ce79a5810414d",
-    "title": "Zapas Converse",
-    "photo": "https://picsum.photos/",
-    "category": "calzado",
-    "price": 81000
-  },
-  {
-    "id": "d416e32e9c377ff4785fcd5c",
-    "title": "Riñonera Dama",
-    "photo": "https://picsum.photos/",
-    "category": "accesorios",
-    "price": 11400,
-    "stock": 9
-  },
-  {
-    "id": "835039e972271f0e49982c5e",
-    "title": "Zapas Virtus",
-    "photo": "https://picsum.photos/",
-    "category": "calzado",
-    "price": 50000,
-    "stock": 5
-  },
-  {
-    "id": "57dacf7b82e858a8105804f0",
-    "title": "Camisa",
-    "photo": "https://picsum.photos/",
-    "category": "ropa masculina",
-    "price": 27000,
-    "stock": 30
-  },
-  {
-    "id": "4adc2ac736bc5268fc36f5b8",
-    "title": "Jeans",
-    "photo": "https://picsum.photos/",
-    "category": "ropa masculina",
-    "price": 21500,
-    "stock": 10
-  },
-  {
-    "id": "a774d146fe088fdcb79db51a",
-    "title": "Buzo",
-    "photo": "https://picsum.photos/",
-    "category": "ropa masculina",
-    "price": 20200,
-    "stock": 6
-  },
-  {
-    "id": "2ee4ec3c3cfabf8012c13561",
-    "title": "remeron",
-    "photo": "https://picsum.photos/",
-    "category": "indumentaria masculina"
-  }
-]
