+++ src/data/fs/files/product.json
@@ -1,0 +1,330 @@
+[
+  {
+    "id": "40fe27e49570ae6626402a3a",
+    "title": "Remera Estampada",
+    "photo": "https://picsum.photos/",
+    "category": "ropa masculina",
+    "price": 8500,
+    "stock": 20
+  },
+  {
+    "id": "d63287b79043ac31ccd6a4a4",
+    "title": "Pantalón de Jean",
+    "photo": "https://picsum.photos/",
+    "category": "ropa masculina",
+    "price": 12500,
+    "stock": 15
+  },
+  {
+    "id": "babd3dc9422c450145721411",
+    "title": "Camisa de Cuadros",
+    "photo": "https://picsum.photos/",
+    "category": "ropa masculina",
+    "price": 9500,
+    "stock": 18
+  },
+  {
+    "id": "bb7576e8d8fd037df09dad35",
+    "title": "Zapatillas Deportivas",
+    "photo": "https://picsum.photos/",
+    "category": "calzado",
+    "price": 15500,
+    "stock": 25
+  },
+  {
+    "id": "e84d356eddea331a98369aca",
+    "title": "Vestido Floral",
+    "photo": "https://picsum.photos/",
+    "category": "ropa femenina",
+    "price": 13500,
+    "stock": 12
+  },
+  {
+    "id": "a9c4b6477f8fbdeb12f83668",
+    "title": "Falda Plisada",
+    "photo": "https://picsum.photos/",
+    "category": "ropa femenina",
+    "price": 10500,
+    "stock": 20
+  },
+  {
+    "id": "f7a2dbeb0fb4a6e1d7756ccf",
+    "title": "Campera de Cuero",
+    "photo": "https://picsum.photos/",
+    "category": "ropa unisex",
+    "price": 28000,
+    "stock": 10
+  },
+  {
+    "id": "ad1d28c9ff75ff1ee8b05ab2",
+    "title": "Bufanda Tejida",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 5500,
+    "stock": 30
+  },
+  {
+    "id": "8c79ca2145ea7b01439ffa57",
+    "title": "Sombrero de Paja",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 7500,
+    "stock": 25
+  },
+  {
+    "id": "3b86e25bcd563c7a95df99db",
+    "title": "Mochila Urbana",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 12500,
+    "stock": 15
+  },
+  {
+    "id": "bf208d39dbbfa97d4f1b0433",
+    "title": "Gorra Deportiva",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 4500,
+    "stock": 35
+  },
+  {
+    "id": "4b2977ec97aa641ce1bdca39",
+    "title": "Reloj Analógico",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 9500,
+    "stock": 20
+  },
+  {
+    "id": "6547b010c440cfa65066ad8e",
+    "title": "Pulsera de Cuero",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 3500,
+    "stock": 40
+  },
+  {
+    "id": "8e49850ebedfb91979db9be9",
+    "title": "Collar de Perlas",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 10500,
+    "stock": 10
+  },
+  {
+    "id": "281d9f8a64f7d2209ed80444",
+    "title": "Lentes de Sol",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 8500,
+    "stock": 25
+  },
+  {
+    "id": "639986883416c970ed252ba4",
+    "title": "Botas de Montaña",
+    "photo": "https://picsum.photos/",
+    "category": "calzado",
+    "price": 19500,
+    "stock": 20
+  },
+  {
+    "id": "f6b02e9c9fe4242af00ab06b",
+    "title": "Sandalias Playeras",
+    "photo": "https://picsum.photos/",
+    "category": "calzado",
+    "price": 9500,
+    "stock": 15
+  },
+  {
+    "id": "833766c1648a63ecf674485f",
+    "title": "Sweater Tejido",
+    "photo": "https://picsum.photos/",
+    "category": "ropa unisex",
+    "price": 11500,
+    "stock": 18
+  },
+  {
+    "id": "4d8431fb10fdf31a96654e35",
+    "title": "Shorts Deportivos",
+    "photo": "https://picsum.photos/",
+    "category": "ropa unisex",
+    "price": 7500,
+    "stock": 22
+  },
+  {
+    "id": "09815d7f28bd125374d55e39",
+    "title": "Top Deportivo",
+    "photo": "https://picsum.photos/",
+    "category": "ropa femenina",
+    "price": 6500,
+    "stock": 30
+  },
+  {
+    "id": "4bff748e88098a012acc7f26",
+    "title": "Traje de Baño",
+    "photo": "https://picsum.photos/",
+    "category": "ropa unisex",
+    "price": 18500,
+    "stock": 10
+  },
+  {
+    "id": "c309161e800b336cf6e42377",
+    "title": "Calcetines Deportivos",
+    "photo": "https://picsum.photos/",
+    "category": "ropa unisex",
+    "price": 2500,
+    "stock": 50
+  },
+  {
+    "id": "947cf1037d8b1e0e3f79d4fa",
+    "title": "Gafas de Natación",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 6500,
+    "stock": 15
+  },
+  {
+    "id": "c746e1b0781183d5d117d7e9",
+    "title": "Paraguas Transparente",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 8500,
+    "stock": 20
+  },
+  {
+    "id": "def4ff8b8cea7af5fd2aa01a",
+    "title": "Chaqueta Impermeable",
+    "photo": "https://picsum.photos/",
+    "category": "ropa unisex",
+    "price": 19500,
+    "stock": 12
+  },
+  {
+    "id": "c9b51e334ac7f9a68ae0f468",
+    "title": "Chaleco Acolchado",
+    "photo": "https://picsum.photos/",
+    "category": "ropa unisex",
+    "price": 13500,
+    "stock": 18
+  },
+  {
+    "id": "646f84ea06ddf2248ea730f5",
+    "title": "Bufanda de Lana",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 6500,
+    "stock": 25
+  },
+  {
+    "id": "561f2ed7cc1a6830385c708c",
+    "title": "Guantes de Invierno",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 4500,
+    "stock": 30
+  },
+  {
+    "id": "7abebfababc02828ffdd3038",
+    "title": "Cinturón de Cuero",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 8500,
+    "stock": 20
+  },
+  {
+    "id": "59f53f57833896ea36270742",
+    "title": "Mochila Escolar",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 11500,
+    "stock": 15
+  },
+  {
+    "id": "4e9165c2ab3547b5f8ab8b02",
+    "title": "Gorro de Lana",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 5500,
+    "stock": 22
+  },
+  {
+    "id": "781ddaafaa983b8ff529cbbd",
+    "title": "Bufanda de Seda",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 10500,
+    "stock": 10
+  },
+  {
+    "id": "d94d080baaaa82d29b050c85",
+    "title": "Bolso de Cuero",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 18500,
+    "stock": 8
+  },
+  {
+    "id": "f47cdb9a39116cb9222667dc",
+    "title": "Mochila de Viaje",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 25500,
+    "stock": 5
+  },
+  {
+    "id": "42f53cbd7296487e7a414a2d",
+    "title": "Maletín Ejecutivo",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 17500,
+    "stock": 10
+  },
+  {
+    "id": "b0acdf504ea03f40a6c3bc3b",
+    "title": "Portafolio de Negocios",
+    "photo": "https://picsum.photos/",
+    "category": "accesorios",
+    "price": 9500,
+    "stock": 15
+  },
+  {
+    "id": "e273f6fdbcfba47fb0de90b0",
+    "title": "Zapatos Formales",
+    "photo": "https://picsum.photos/",
+    "category": "calzado",
+    "price": 18500,
+    "stock": 12
+  },
+  {
+    "id": "a78cb4fea5cf556d41368725",
+    "title": "Botas de Cuero",
+    "photo": "https://picsum.photos/",
+    "category": "calzado",
+    "price": 24500,
+    "stock": 10
+  },
+  {
+    "id": "0c386d617e224956798d572a",
+    "title": "Alpargatas",
+    "photo": "https://picsum.photos/",
+    "category": "calzado",
+    "price": 6500,
+    "stock": 20
+  },
+  {
+    "id": "2b5174589c4b0a2e9d06a6c4",
+    "title": "Zapatillas Urbanas",
+    "photo": "https://picsum.photos/",
+    "category": "calzado",
+    "price": 10500,
+    "stock": 15
+  },
+  {
+    "id": "a64162f2cb06b06a5c8ce6d9",
+    "title": "Mocasines de Cuero",
+    "photo": "https://picsum.photos/",
+    "category": "calzado",
+    "price": 12500,
+    "stock": 18
+  }
+]
