+++ src/data/fs/files/user.json
@@ -1,30 +1,30 @@
 [
   {
-    "id": "782613ed2274e5443811a7b8",
+    "id": "4fdaea51f2ab877ffd598de2",
     "photo": "https://picsum.photos/",
     "email": "email1@gmail.com",
     "password": "pass1",
-    "role": 5
+    "role": 0
   },
   {
-    "id": "11180dddf7e6f2b491a3b422",
+    "id": "0853c2fe648bb1e66cdf8773",
     "photo": "https://picsum.photos/",
     "email": "email2@gmail.com",
+    "password": "pass1",
+    "role": 0
+  },
+  {
+    "id": "a07ec5577d9260912287f1fc",
+    "photo": "https://picsum.photos/",
+    "email": "email3@gmail.com",
     "password": "pass1",
     "role": 1
   },
   {
-    "id": "8c92b6c199d5d1e3538f98ea",
-    "photo": "https://picsum.photos/",
-    "email": "email3@gmail.com",
-    "password": "pass1",
-    "role": 5
-  },
-  {
-    "id": "a10cf938ffc9bd211143b10c",
+    "id": "d47727475f779c1497af94d6",
     "photo": "https://picsum.photos/",
     "email": "email4@gmail.com",
     "password": "pass1",
-    "role": 5
+    "role": 1
   }
 ]
